Displaying color frames.
Former-commit-id: cc1926cf16caf4672768ad7790c8c2621191f453
Former-commit-id: f28485b1f26a920549e47efcbc2dcdc99181006d

+++ VolumeCheckerGUI/Utils/Constants.cs
@@ -6,6 +6,6 @@
 
 		public const float FovY = 57; 
 
-		public const int MinDepth = 300;
+		public const short MinDepth = 300;
 	}
 }
